Refactor FuzzyMatch to its own file & add testing
Added setup for unit testing using the new
[Swift Testing](https://developer.apple.com/xcode/swift-testing/)
framwork. Added it to the test pipeline.

New `make run` command since `make test` now runs the test suit.

Fixes #22
