+++ .github/workflows/test.yml
@@ -18,3 +18,6 @@ jobs:
 
       - name: Build
         run: make build
+
+      - name: Test
+        run: make test
