+++ Makefile
@@ -3,16 +3,18 @@ BINARY_NAME = mac-menu
 SRC_DIR = src
 DIST_DIR = .build
 VERSION ?= $(shell git describe --tags --always --dirty 2>/dev/null || echo "unknown")
+SOURCES = $(SRC_DIR)/FuzzyMatch.swift $(SRC_DIR)/main.swift $(SRC_DIR)/Version.swift
 
-.PHONY: all clean install uninstall test
+.PHONY: all clean install uninstall test run
+RUN_CMD = $(DIST_DIR)/$(BINARY_NAME)
 
 all: $(DIST_DIR)/$(BINARY_NAME)
 
 build: all
 
-$(DIST_DIR)/$(BINARY_NAME): $(SRC_DIR)/main.swift $(SRC_DIR)/Version.swift
+$(DIST_DIR)/$(BINARY_NAME): $(SOURCES)
 	@mkdir -p $(DIST_DIR)
-	swiftc -O -o $(DIST_DIR)/$(BINARY_NAME) $(SRC_DIR)/main.swift $(SRC_DIR)/Version.swift -framework Cocoa
+	swiftc -O -o $(DIST_DIR)/$(BINARY_NAME) $(SOURCES) -framework Cocoa
 
 $(SRC_DIR)/Version.swift:
 	@echo 'let appVersion = "$(VERSION)"' > $(SRC_DIR)/Version.swift
@@ -31,6 +33,9 @@ clean:
 	@echo "Cleaning build files..."
 	@rm -rf $(DIST_DIR)
 
-test: $(DIST_DIR)/$(BINARY_NAME)
-	@echo "Testing with Yes/No input..."
-	@echo "Yes\nNo" | $(DIST_DIR)/$(BINARY_NAME) 
+test:
+	@echo "Running Swift Testing suite..."
+	@swift test --enable-swift-testing
+
+run: build
+	@$(RUN_CMD)
